非0退出时也自动重启

+++ restarter.py
@@ -15,18 +15,24 @@ def run(cmd_file, restart_interval):
     """ 运行接口 """
     while True:
         print datetime.now().strftime('%b-%d-%y %H:%M:%S')
+
         process = Popen(cmd_file)
         for _ in range(restart_interval):
             sleep(1)
             # 如果进程正常结束，则无需再次重启
-            sig = process.poll()
+            sig = process.poll()            
             if not sig is None:
-                return
-        system('taskkill /F /T /PID ' + str(process.pid))
+                if process.returncode == 0:
+                    return
+                else:
+                    break
 
-        # 等待一段时间，避免下一次启动时由于之前的资源占用导致进程启动失败
-        # 譬如cmd_file中包含输出到文件的操作
-        sleep(1)
+        if process.poll() is None:
+            system('taskkill /F /T /PID ' + str(process.pid))
+
+            # 等待一段时间，避免下一次启动时由于之前的资源占用导致进程启动失败
+            # 譬如cmd_file中包含输出到文件的操作
+            sleep(1)
 
 
 if __name__ == '__main__':
